Fix args with whitespace are not passed to php properly.
__Problem__
Running PHP and passing an argument with whitespace does not work.
For example, the following command:

php -r "echo 'boop!';"

Results in the following error:

PHP Parse error:  syntax error, unexpected end of file in Command line code on line 1

__Cause__
This is because the double quotes around the argument are removed by the shell
when it is passed to the php script. The script them passes the arguments to the
container by using the $@ variable. This results in the following being executed:

php -r echo 'boop!';

__Fix__
I modified the script to iterate over the arguments in $@ and check if any contain
whitespace. If they do, they are wrapped in double quotes. If not, they are used
as-is.