Fix 7.1 packages no longer available.
Problem:
The image is failing to build, complaining that the PHP 7.1 packages are not available in the repos. This image previously built fine, so the packages seem to have been removed.

Cause:
According to [launchpad](https://launchpad.net), [php7.1-cli](https://launchpad.net/ubuntu/zesty/+package/php7.1-cli) for amd64 is in "Release", but the [release index](http://archive.ubuntu.com/ubuntu/indices/override.zesty.universe) does not have it listed. However, it is listed in the [proposed index](http://archive.ubuntu.com/ubuntu/indices/override.zesty-proposed.universe). I am not sure why or when the packages were removed from the release universe repo.

Fix:
Since Zesty is currently in a "devel" state, and is therefore subject to these breaking changes, I have switched back to the current LTS ubuntu release yakkety, and switched to use the ondrej PPA for the 7.1 installation.