Fix user ID 1000 is assumed
Problem:
The container was assuming that volumes mounted would be owned by user id 1000. This is sometimes true, but is not always the case.

Fix:
I modified the container to accept environmental variables to define what user id the initial process and/or php-fpm process should run as. See README.md for details.