Add --no-install-recommends
Additional recommended packages are being installed that are not strictly necessary to meet the requirements of PHP 7 or Laravel.

I have removed these packages to decrease the Docker image size, and will leave it up to the extending packages to install any of these requirements that are needed for their particular use-cases.